Basic UI
This is the basic menu UI created for the game. It includes a title screen, instructions, an inventory screen, and a pause menu. The appearance is not final. I've also had to do some programming to bring some functionality to the UI. As a side effect of programming the Pause UI, pausing is implemented.