Game Over Screen
I've got the game over screen programmed. As long as the "Testing Build" boolean is checked off, a Game Over can be toggled instantly even without a player in my branch. The game over screen starts with a black circle expanding across the screen followed by the text appearing. The art gradually fades in after that. When the art is fully loaded, the buttons appear. They are fully functional. The art is not final.